chore(connlib): extract flow tracker into dedicated crate
