+++ rust/libs/connlib/tunnel/src/tests/sut.rs
@@ -1380,7 +1380,7 @@ fn address_from_destination(
 }
 
 fn test_ingest_token() -> crate::messages::IngestToken {
-    serde_json::from_str(&format!("\"{}\"", crate::messages::TEST_INGEST_TOKEN)).unwrap()
+    serde_json::from_str(&format!("\"{}\"", flow_tracker::TEST_INGEST_TOKEN)).unwrap()
 }
 
 fn make_preshared_key_and_ice(
